Fix scheduler errors with larger MAX_TASKS.

+++ stm32/include/trf_scheduler.h
@@ -5,8 +5,8 @@
 
 #define MAX_TASKS 256
 
-uint8_t SCH_AddTask(void (*func)(), uint32_t delay, uint32_t period);
-void SCH_RemoveTask(uint8_t index);
+uint32_t SCH_AddTask(void (*func)(), uint32_t delay, uint32_t period);
+void SCH_RemoveTask(uint32_t index);
 void SCH_DispatchTasks(void);
 void SCH_Tick(void);
